requirements.txt
flask
torch
torchvision
facenet-pytorch
opencv-python
numpy
Pillow
